move together rest into together
